(1/2) Make ROM zip compatible with ext4/f2fs
Need a couple things to get this working:

First, you need to use the "f2fs2" branch of the android_build project. If building for another ROM, just look at the branch and it will be pretty easy to tell which commits you need from it.

Second,you need this commit in system/core: http://gerrit.liquidsmooth.net/#/c/199/ (just in case we don't end up merging it into the main branch in the event that it breaks non-f2fs builds).

Lastly, you obviously need f2fs drivers built and enabled in your kernel if not using ours.

Change-Id: I8c8213cea0a0a66648073d63b3c2dd1595ab2157